HOTT-3641: Extract shared component behaviour into concern
