+++ app/models/measure_condition_component.rb
@@ -1,5 +1,4 @@
 class MeasureConditionComponent < Sequel::Model
-  plugin :time_machine
   plugin :oplog, primary_key: %i[measure_condition_sid
                                  duty_expression_id]
 
@@ -8,38 +7,7 @@ class MeasureConditionComponent < Sequel::Model
   one_to_one :measure_condition, key: :measure_condition_sid,
                                  primary_key: :measure_condition_sid
 
-  one_to_one :duty_expression, key: :duty_expression_id,
-                               primary_key: :duty_expression_id do |ds|
-    ds.with_actual(DutyExpression)
-  end
-
-  one_to_one :measurement_unit, key: :measurement_unit_code,
-                                primary_key: :measurement_unit_code do |ds|
-    ds.with_actual(MeasurementUnit)
-  end
-
-  one_to_one :monetary_unit, key: :monetary_unit_code,
-                             primary_key: :monetary_unit_code do |ds|
-    ds.with_actual(MonetaryUnit)
-  end
-
-  one_to_one :measurement_unit_qualifier, key: :measurement_unit_qualifier_code,
-                                          primary_key: :measurement_unit_qualifier_code do |ds|
-    ds.with_actual(MeasurementUnitQualifier)
-  end
-
-  delegate :description, :abbreviation, to: :duty_expression, prefix: true
-  delegate :abbreviation, to: :monetary_unit, prefix: true, allow_nil: true
-  delegate :description, to: :monetary_unit, prefix: true, allow_nil: true
-
-  def ad_valorem?
-    monetary_unit_code.nil? &&
-      measurement_unit_code.nil?
-  end
-
-  def zero_duty?
-    duty_amount&.zero?
-  end
+  include Componentable
 
   def formatted_duty_expression
     DutyExpressionFormatter.format(
@@ -54,16 +22,5 @@ class MeasureConditionComponent < Sequel::Model
       currency: TradeTariffBackend.currency,
       formatted: true,
     )
-  end
-
-  def expresses_unit?
-    measurement_unit_code
-  end
-
-  def unit
-    {
-      measurement_unit_code: measurement_unit_code,
-      measurement_unit_qualifier_code: measurement_unit_qualifier_code,
-    }
   end
 end
